==> 1.Getting_files.ipynb <==
import os
from eyed3.mp3 import Mp3AudioFile
%config IPCompleter.greedy=True
import pandas as pd

df = pd.DataFrame()
for roots, dirs, files in os.walk("/media/grzegorz/936e354d-72a9-4039-c5a7-9e550f310451/spotify_search/music"):
    for file in files:
        path = os.path.join(roots, file)
        audiofile = Mp3AudioFile(path)
        data = []
        try:
            data.append(audiofile.tag.artist)
            data.append(audiofile.tag.title)
        except:
            pass
        df = df.append(pd.Series(data), ignore_index=True)

df = df.iloc[:,2:]
print(df)

df.to_csv('my_music_database.csv')
