==> 1.new.get_albums_data.ipynb <==
import pandas as pd
import time
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials

io = './eksporty/int72.xlsx'
df = pd.read_excel(io).iloc[:,:1]
df2 = df['Title'].str.split(pat=' – ',expand=True)
df2[1] = df2[1].str[:-6]

df['Artist'] = df2[0]
df['Album_title'] = df2[1]
df.head()

%run -i '1. ids.py'

columns = ['artist','album','track_title','track_title','track_id']
output_df = pd.DataFrame(columns=columns)
output_df = output_df.fillna(0)

client_credentials_manager = SpotifyClientCredentials(client_id=client_id, client_secret=client_secret)
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)
for x in range(len(df)):
    artist= df.iloc[x,1]
    album= df.iloc[x,2]

    album_id = sp.search(q="" + artist + " " + album + "", type='album', limit=1)
    try:
        album_id = album_id['albums']['items'][0]['uri']
    except:
        pass
    try:
        tracks = sp.album_tracks(album_id)
    except:
        pass
    for track in tracks['items']:
        track_title = track['name']
        track_nr = track['track_number']
        track_id = track['id']
        temp_df = pd.DataFrame([[artist, album, track_title, track_nr, track_id]],columns=columns)
        output_df = output_df.append(temp_df, ignore_index=True)

output_df

output_df.to_csv('albums_data_with_ids.csv')
