==> 2.a.new.get_albums_props.ipynb <==
import pandas as pd
import time
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials


df = pd.read_csv('albums_with_props.csv').iloc[:,1:]
df = df.dropna()
df.isna().sum()

df2 = df.copy()

%run -i '1. ids.py'


client_credentials_manager = SpotifyClientCredentials(client_id=client_id, client_secret=client_secret)
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)


for x in range(len(df2.index)):
    ids= df2.iloc[x,4]
    track_data = sp.audio_analysis(ids)
    if track_data != [None]:
        
        bars = 0
        for item in track_data['bars']:
             bars += 1
        df2.loc[x, 'bars'] = bars
        sections = 0
        for item in track_data['sections']:
            sections += 1
        df2.loc[x, 'sections'] = sections
        segments = 0
        for item in track_data['segments']:
            segments += 1
        df2.loc[x, 'segments'] = segments    

        if not x % 50:
            print(x)
            time.sleep(2) 


df2 = df2.dropna()

print(df2.isna().sum())


df2.to_csv('albums_with_props.csv')

df2.head()

df2.tail()
