==> 2.get_data_to_train.ipynb <==
import pandas as pd
import time
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials


df = pd.read_csv('my_music_database.csv').iloc[:,1:]
df = df.dropna()
df.shape


%run -i '1. ids.py'


client_credentials_manager = SpotifyClientCredentials(client_id=client_id, client_secret=client_secret)
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)


for x in range(len(df.index)):
    artist= df.iloc[x,0]
    track= df.iloc[x,1]

    track_id = sp.search(q="" + artist + " " + track + "", type='track', limit=1)
    for item in track_id['tracks']['items']:
        df.loc[x,'id'] = item['id']
    if not x % 50:
        print(x)
        time.sleep(2)
        print(x)

x

df.to_csv('data_with_ids.csv')

df.head()

df.tail()

df.isna().sum()
