==> 2.new.get_albums_props.ipynb <==
import pandas as pd
import time
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials


df = pd.read_csv('albums_data_with_ids.csv').iloc[:,1:]
df = df.dropna()
df.isna().sum()

df2 = df.copy()

%run -i '1. ids.py'


client_credentials_manager = SpotifyClientCredentials(client_id=client_id, client_secret=client_secret)
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)


for x in range(len(df2.index)):
    ids= df2.iloc[x,4]
    track_id = sp.audio_features(ids)
    if track_id != [None]:
        
        for item in track_id:
            df2.loc[x, 'id2' ]= item['id']
            df2.loc[x, 'danceability'] = item['danceability']
            df2.loc[x, 'energy'] = item['energy']
            df2.loc[x, 'key'] = item['key']
            df2.loc[x, 'loudness'] = item['loudness']
            df2.loc[x, 'mode'] = item['mode']
            df2.loc[x, 'speechiness'] = item['speechiness']
            df2.loc[x, 'acousticness'] = item['acousticness']
            df2.loc[x, 'instrumentalness'] = item['instrumentalness']
            df2.loc[x, 'liveness'] = item['liveness']
            df2.loc[x, 'valence'] = item['valence']
            df2.loc[x, 'tempo'] = item['tempo']
            df2.loc[x, 'type'] = item['type']
            df2.loc[x, 'type'] = item['type']
            df2.loc[x, 'duration_ms'] = item['duration_ms']
            df2.loc[x, 'time_signature'] = item['time_signature']

       
        if not x % 50:
            print(x)
            time.sleep(2)
            


df2 = df2.dropna()

print(df2.isna().sum())


df2.to_csv('albums_with_props.csv')

df2.head()

df2.tail()
