==> 99.Predict.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split, ParameterGrid
import collections


from sklearn import svm
from sklearn.neighbors import LocalOutlierFactor
from sklearn.ensemble import IsolationForest

X = pd.read_csv('data_with_props.csv').iloc[:,5:]
X = X.drop('type', axis=1)
X_out = pd.read_csv('albums_with_props.csv').iloc[:,7:]
X_out = X_out.drop('type', axis=1)


X_train, X_test = train_test_split(X, test_size=0.1, random_state=42)

def custom_scoring(clf, X_out, X_train):
    
    y_pred_train = clf.predict(X_train)
    y_pred_out = clf.predict(X_out)
    

    a = collections.Counter(y_pred_train)
    b = collections.Counter(y_pred_out)
    print(a)
    print(b)
    try:
        train = dict(a.items())[1]/(dict(a.items())[-1] + dict(a.items())[1])
        outside = dict(b.items())[-1]/(dict(b.items())[1] + dict(b.items())[-1])
        print(train)
        print(outside)
    except:
        train = 0
        
    if train > 0.8 and outside <0.8:
        result = outside
        print(result)
    else:
        result = 0

        
    return(result)

def fit_and_display(grid, clf, X_out, X_train):
    best_score_list = [0]
    best_params_list = []
    for g in ParameterGrid(grid):
        clf.set_params(**g)
        clf.fit(X_train)
        score = custom_scoring(clf, X_out, X_train)
        if score > max(best_score_list):
            
            best_score_list.append(score)
            params = clf.get_params(deep=True)
            best_params_list.append(clf.get_params(params))
            print("")
            print('***********************')
            print('Best at the moment!')
            print(score)
            print(clf.get_params(deep=True))
            print('***********************')
            print("")
    return(best_score_list, best_params_list)    

def check_best_result(clf, best_params_list, X_train, X_test, X_out):
    clf.set_params(**best_params_list[-1])

    y_pred_train = clf.predict(X_train)
    y_pred_test = clf.predict(X_test)
    y_pred_out = clf.predict(X_out)
    print('Values for train, test and out:')
    a = (collections.Counter(y_pred_train))
    b = (collections.Counter(y_pred_test))
    c = (collections.Counter(y_pred_out))
    print('Scoring for train, test and out')
    train = dict(a.items())[1]/(dict(a.items())[-1] + dict(a.items())[1])
    test = dict(b.items())[1]/(dict(b.items())[-1] + dict(b.items())[1])
    outside = dict(c.items())[-1]/(dict(c.items())[1] + dict(c.items())[-1])
    print(train, test, outside)
    print('Parameters of the best estimator:')
    print(clf.get_params)   

kernels = [ 'poly', 'rbf', 'sigmoid', 'linear']
degree = [2,3,4,5,6]
gammas = [0.001, 0.01, 0.1, 1]
nus = [0.01, 0.1, 0.15, 0.2, 0.3, 0.4, 0.5]
tols = [0.001, 0.01, 0.1]
grid = {'degree': degree, 'gamma' : gammas, 'kernel' : kernels, 'nu' : nus, 'tol' : tols}
clf = svm.OneClassSVM(verbose=10)

best_score_list, best_params_list = fit_and_display(grid, clf, X_out, X_train)

check_best_result(clf, best_params_list, X_train, X_test, X_out)

leafs = [ 1,3,5,9,12,17,20,25,35,50]
neighbours = [ 1,3,5,9,12,17,20,25,35,50]
contaminations = [0.001, 0.01, 0.15,0.2,0.5]
p = [1,2]
algs = ['auto', 'ball_tree', 'kd_tree', 'brute']

grid = {'leaf_size': leafs, 'n_neighbors' : neighbours, 'contamination' : contaminations, 'algorithm' : algs, 'p': p}
clf = LocalOutlierFactor(novelty=True, n_jobs=-1)

best_score_list, best_params_list = fit_and_display(grid, clf, X_out, X_train)

check_best_result(clf, best_params_list, X_train, X_test, X_out)

estims = [ 20,50,80,120,180,400,600,900,1300,2000,3000,5000,8000]
neighbours = [ 1,3,5,9,12,17,20,25,35,50]
contaminations = [0.001, 0.01, 0.15,0.2,0.5]
boots = [True, False]

grid = { 'n_estimators' : estims, 'contamination' : contaminations, 'bootstrap' : boots}
clf = IsolationForest(behaviour="new", n_jobs=-1, verbose=10)

best_score_list, best_params_list = fit_and_display(grid, clf, X_out, X_train)

check_best_result(clf, best_params_list, X_train, X_test, X_out)
